Update _fit_inductive_matrix_completion as _fit_imc
The update of the _fit_imc method improves naming as well as allowing the option of solving for
both W and H simultaneously or just W (if H is already known, used for the transform method).
In addition the method allows the user to choose the alorithm used by the minimization function.