Add _check_x method
Using this class with crossvalidation in sklearn requires that all data be packaged
in a single object (i.e. X instead of X, Y). To facilitate this we will create a
helper class that will allow the data to be validated and split by the cross
validation generators returning a tuple of (X, Y) for use in the fit and score
methods.