Remove implementation using W and H
The original implementation utilized alternating minimization of W and H arrays. While
this may be useful in a case with very large numbers of features we decided to go with
a more performant method focusing on just a single array Z. This array can be decomposed
into W and H by svd, nmf or other methods.