Update score method and remove fit transform
It became apparent that a fit_transform method was not the best approach for this model
(possibility of too large final matrix and redundance with predict_all method). The
score method was updated to use the check_is_fitted method and take an array of user
and item ids, and an array of true values as input.