Make improvements to calculation of dot product diagonals
The previous method for calculating the diagnol of the dot product of two arrays
was found to be less efficient and more verbose than another method. Updated any
instance of the old method to the new method.

For reference the new faster method for calculating diagonals takes the elementwise
multiplication of an array with the transpose of the second array, then sums over
the last dimension.