Created dimc.py
dimc.py holds the DirtyIMC class. This is a method that combines and IMC
method with another standard matrix completion method to help minimize error.

For now the implementation uses IMC and ALS defined in this package, but
in theory any method that uses side information could be combined with any
method of direct matrix completion.