Implement _update_parallel with joblib
Using joblib.Parallel I removed the use of ProcessPoolExecutor. This modification
has also allowed me to bring the methods for updating the user and item features
back into the class as joblib.Prallel manages the memory issues that lead me to
run the code in a separate main thread originally.